requirement.txt
streamlit
opencv-python-headless
mediapipe
numpy
scikit-learn
joblib
Pillow
